refactor(indexeddb): deduplicate serializer traits and types
