+++ crates/matrix-sdk-indexeddb/src/serializer/indexed_type/mod.rs
@@ -1,0 +1,156 @@
+// Copyright 2025 The Matrix.org Foundation C.I.C.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License
+
+pub mod range;
+pub mod traits;
+
+use gloo_utils::format::JsValueSerdeExt;
+use range::IndexedKeyRange;
+use serde::{de::DeserializeOwned, Serialize};
+use thiserror::Error;
+use traits::{Indexed, IndexedKey};
+use wasm_bindgen::JsValue;
+use web_sys::IdbKeyRange;
+
+use crate::serializer::SafeEncodeSerializer;
+
+#[derive(Debug, Error)]
+pub enum IndexedTypeSerializerError<IndexingError> {
+    #[error("indexing: {0}")]
+    Indexing(IndexingError),
+    #[error("serialization: {0}")]
+    Serialization(#[from] serde_json::Error),
+}
+
+impl<T> From<serde_wasm_bindgen::Error> for IndexedTypeSerializerError<T> {
+    fn from(e: serde_wasm_bindgen::Error) -> Self {
+        Self::Serialization(serde::de::Error::custom(e.to_string()))
+    }
+}
+
+/// A (de)serializer for an IndexedDB implementation of [`EventCacheStore`][1].
+///
+/// This is primarily a wrapper around [`SafeEncodeSerializer`] with
+/// convenience functions for (de)serializing types specific to the
+/// [`EventCacheStore`][1].
+///
+/// [1]: matrix_sdk_base::event_cache::store::EventCacheStore
+#[derive(Debug, Clone)]
+pub struct IndexedTypeSerializer {
+    inner: SafeEncodeSerializer,
+}
+
+impl IndexedTypeSerializer {
+    pub fn new(inner: SafeEncodeSerializer) -> Self {
+        Self { inner }
+    }
+
+    /// Returns a reference to the inner [`IndexeddbSerializer`].
+    pub fn inner(&self) -> &SafeEncodeSerializer {
+        &self.inner
+    }
+
+    /// Encodes an key for a [`Indexed`] type.
+    ///
+    /// Note that the particular key which is encoded is defined by the type
+    /// `K`.
+    pub fn encode_key<T, K>(&self, components: K::KeyComponents<'_>) -> K
+    where
+        T: Indexed,
+        K: IndexedKey<T>,
+    {
+        K::encode(components, &self.inner)
+    }
+
+    /// Encodes a key for a [`Indexed`] type as a [`JsValue`].
+    ///
+    /// Note that the particular key which is encoded is defined by the type
+    /// `K`.
+    pub fn encode_key_as_value<T, K>(
+        &self,
+        components: K::KeyComponents<'_>,
+    ) -> Result<JsValue, serde_wasm_bindgen::Error>
+    where
+        T: Indexed,
+        K: IndexedKey<T> + Serialize,
+    {
+        serde_wasm_bindgen::to_value(&self.encode_key::<T, K>(components))
+    }
+
+    /// Encodes a key component range for an [`Indexed`] type.
+    ///
+    /// Note that the particular key which is encoded is defined by the type
+    /// `K`.
+    pub fn encode_key_range<T, K>(
+        &self,
+        range: impl Into<IndexedKeyRange<K>>,
+    ) -> Result<IdbKeyRange, serde_wasm_bindgen::Error>
+    where
+        T: Indexed,
+        K: Serialize,
+    {
+        use serde_wasm_bindgen::to_value;
+        Ok(match range.into() {
+            IndexedKeyRange::Only(key) => IdbKeyRange::only(&to_value(&key)?)?,
+            IndexedKeyRange::Bound(lower, upper) => {
+                IdbKeyRange::bound(&to_value(&lower)?, &to_value(&upper)?)?
+            }
+        })
+    }
+
+    /// Encodes a key component range for an [`Indexed`] type.
+    ///
+    /// Note that the particular key which is encoded is defined by the type
+    /// `K`.
+    pub fn encode_key_component_range<'a, T, K>(
+        &self,
+        range: impl Into<IndexedKeyRange<K::KeyComponents<'a>>>,
+    ) -> Result<IdbKeyRange, serde_wasm_bindgen::Error>
+    where
+        T: Indexed,
+        K: IndexedKey<T> + Serialize,
+    {
+        let range = match range.into() {
+            IndexedKeyRange::Only(components) => {
+                IndexedKeyRange::Only(K::encode(components, &self.inner))
+            }
+            IndexedKeyRange::Bound(lower, upper) => {
+                let lower = K::encode(lower, &self.inner);
+                let upper = K::encode(upper, &self.inner);
+                IndexedKeyRange::Bound(lower, upper)
+            }
+        };
+        self.encode_key_range::<T, K>(range)
+    }
+
+    /// Serializes an [`Indexed`] type into a [`JsValue`]
+    pub fn serialize<T>(&self, t: &T) -> Result<JsValue, IndexedTypeSerializerError<T::Error>>
+    where
+        T: Indexed,
+        T::IndexedType: Serialize,
+    {
+        let indexed = t.to_indexed(&self.inner).map_err(IndexedTypeSerializerError::Indexing)?;
+        serde_wasm_bindgen::to_value(&indexed).map_err(Into::into)
+    }
+
+    /// Deserializes an [`Indexed`] type from a [`JsValue`]
+    pub fn deserialize<T>(&self, value: JsValue) -> Result<T, IndexedTypeSerializerError<T::Error>>
+    where
+        T: Indexed,
+        T::IndexedType: DeserializeOwned,
+    {
+        let indexed: T::IndexedType = value.into_serde()?;
+        T::from_indexed(indexed, &self.inner).map_err(IndexedTypeSerializerError::Indexing)
+    }
+}
