+++ crates/matrix-sdk-indexeddb/src/serializer/indexed_type/traits.rs
